Update package version and implement auto-release workflow
The package version in package.json has been updated from 1.0.0 to 1.0.1. Moreover, an auto-release workflow has been added to the GitHub Action to automate the release process. This workflow will draft the next release notes as Pull Requests are merged into "main" and it also includes an autolabeler feature to help categorize the changes.